Remove a hard dependency on the Unix module
This removes an exception printer which requires Unix to be linked
to use Sexp.  Without this, there is only a "soft" dependency in
the form of Bigarray, which doesn't require unix.cmxa to actually
be present.

It would be great if this exception printer could be moved into a
sexplib.unix subpackage, to make it easier to portably use it
in Mirage or JavaScript.